requirements.txt
influxdb
flask
MarkupSafe
python-dateutil
Markdown
flask-login
flask-sqlalchemy
flask-wtf
wtforms